% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add_labels}
\alias{add_labels}
\title{Generates text labels to go at the end of a line plot, one label per breakdown
group,}
\usage{
add_labels(data, x, y, breakdown, gap_factor = 0.045)
}
\arguments{
\item{data}{}

\item{x}{}

\item{y}{}

\item{breakdown}{}

\item{gap_factor}{}
}
\value{

}
\description{
Generates text labels to go at the end of a line plot, one label per breakdown
group,
}
\examples{



}
