% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{formula_to_sym}
\alias{formula_to_sym}
\title{Turn a formula into a symbol}
\usage{
formula_to_sym(x)
}
\arguments{
\item{x}{a formula that would be passed to plotly functions.  #'}
}
\description{
Turn a formula into a symbol
}
\details{
Turn a formula object (as used in \code{plotly} functions for column names)
into a symbol object (as used in \code{dplyr} functions for column names).
}
\examples{
# Useful for converting between plotly code and dplyr code
data <- data.frame(
 group = "A"
 x = 1:100,
 y = rnorm(100)
)

x_plot <- ~x
y_plot <- ~y

plotly::plot_ly(
  data,
  x = x_plot,
  y = y_plot
)

x_dplyr <- formula_to_sym(x_plot)
y_dplyr <- formula_to_sym(y_plot)

select(data, {{ x_dplyr }}, {{ y_dplyr }})
}
