% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_colour_palette}
\alias{make_colour_palette}
\title{Make a colour palette of \code{n} colours.}
\usage{
make_colour_palette(n, name)
}
\arguments{
\item{n}{Number of difference colours in the palette. Any positive integer.}

\item{name}{ColorBrewer palette name - see \code{\link[RColorBrewer]{RColorBrewer}}
and \href{http://www.colorbrewer.org}{ColorBrewer}}
}
\description{
Works like \code{RcolorBrewer::brewer.pal} but always returns exactly as many
colours as requested.
}
\details{
Uses \code{grDevices::colorRamp} to produce large colour palettes (more than
9 colours) by interpolating ones returned by \code{RColorBrewer::brewer.pal}.
Because of this very large colour palettes will be difficult to distinguish.
}
\examples{
# Make a colour palette with exactly 2 colours
colours <- make_colour_palette(2, "PuBu")
length(colours) == 2
# Compare to RColorBrewer's minimum of 3 colours
colours_brewer <- RColorBrewer::brewer.pal(2, "PuBu")
length(colours_brewer) == 2

# Making palettes with more than 9 colours
colours <- make_colour_palette(11, "PuBu")
length(colours) == 11
colours_brewer <- RColorBrewer::brewer.pal(11, "PuBu")
length(colours_brewer) == 11
}
