% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subplots.R
\name{make_subplots}
\alias{make_subplots}
\title{Title}
\usage{
make_subplots(
  data,
  group_column,
  nrows = 2,
  margin = 0.07,
  shareY = TRUE,
  shareX = TRUE,
  plot_title = NULL,
  y_axis_title = NULL,
  ...
)
}
\arguments{
\item{data}{}

\item{group_column}{}

\item{nrows}{}

\item{margin}{}

\item{shareY}{}

\item{shareX}{}

\item{...}{}

\item{master_y_axis_title}{}
}
\value{

}
\description{
Title
}
\examples{

}
